requirements.txt
python-dotenv
langchain
langchain-community
openai
langgraph
azure-ai-documentintelligence
pypdf